AdaptiveImpute

An implementation of the AdaptiveImpute algorithm for matrix completion for sparse matrices.

adaptive_impute(
  X,
  rank,
  ...,
  initialization = c("svd", "adaptive-initialize", "approximate"),
  max_iter = 200L,
  check_interval = 1L,
  epsilon = 1e-07,
  additional = NULL
)

# S3 method for sparseMatrix
adaptive_impute(
  X,
  rank,
  ...,
  initialization = c("svd", "adaptive-initialize", "approximate"),
  additional = NULL
)

# S3 method for LRMF
adaptive_impute(
  X,
  rank,
  ...,
  epsilon = 1e-07,
  max_iter = 200L,
  check_interval = 1L
)

Arguments

X
A sparse matrix of Matrix::sparseMatrix() class.

rank
Desired rank (integer) to use in the low rank approximation. Must be at least 2L and at most the rank of X. Note that the rank of X is typically unobserved and computations may be unstable or even fail when rank is near or exceeds this threshold.

...
Unused additional arguments.

initialization
How to initialize the low rank approximation. Options are:

"svd" (default). In the initialization step, this treats unobserved values as zeroes.

"adaptive-initialize". In the initialization step, this treats unobserved values as actually unobserved. However, the current AdaptiveInitialize implementation relies on dense matrix computations that are only suitable for relatively small matrices.

"approximate". An approximate variant of AdaptiveInitialize that is less computationally expensive. See adaptive_initialize for details.

Note that initialization matters as AdaptiveImpute optimizes a non-convex objective. The current theory shows that initializing with AdaptiveInitialize leads to a consistent estimator, but it isn't know if this is the case for SVD initialization. Empirically we have found that SVD initialization works well nonetheless.

max_iter
Maximum number of iterations to perform (integer). Defaults to 200L. In practice 10 or so iterations will get you a decent approximation to use in exploratory analysis, and and 50-100 will get you most of the way to convergence. Must be at least 1L.

check_interval
Integer specifying how often to perform convergence checks. Defaults to 1L. In practice, check for convergence requires a norm calculation that is expensive for large matrices and decreasing the frequency of convergence checks will reduce computation time. Can also be set to NULL, which case max_iter iterations of the algorithm will occur with no possibility of stopping due to small relative change in the imputed matrix. In this case delta will be reported as Inf.

epsilon
Convergence criteria, measured in terms of relative change in Frobenius norm of the full imputed matrix. Defaults to 1e-7.

additional
Ignored except when alpha_method = "approximate" in which case it controls the precise of the approximation to alpha. The approximate computation of alpha will always understand alpha, but the approximation will be better for larger values of additional. We recommend making additional as large as computationally tolerable.

Value

A low rank matrix factorization represented by an adaptive_imputation() object.

Examples

### SVD initialization (default) --------------------------------------------

mf <- adaptive_impute(ml100k, rank = 3L, max_iter = 20L)

#> INFO [2020-10-20 22:58:41] Using svd initialization.
#> INFO [2020-10-20 22:58:41] Done initializing.
#> INFO [2020-10-20 22:58:41] Beginning AdaptiveImpute (max 20 iterations).
#> INFO [2020-10-20 22:58:41] Checking convergence every 1 iteration(s).
#> INFO [2020-10-20 22:58:41] Iter 1 complete. delta = 0.22288584976601, alpha = 315.498594765073
#> INFO [2020-10-20 22:58:41] Iter 2 complete. delta = 0.0515957111056737, alpha = 236.47843078421
#> INFO [2020-10-20 22:58:41] Iter 3 complete. delta = 0.0212652804266644, alpha = 192.407973190373
#> INFO [2020-10-20 22:58:42] Iter 4 complete. delta = 0.0111504246107586, alpha = 164.221353309024
#> INFO [2020-10-20 22:58:42] Iter 5 complete. delta = 0.00669435704217216, alpha = 144.663327182091
#> INFO [2020-10-20 22:58:42] Iter 6 complete. delta = 0.00438783102912924, alpha = 130.322069276523
#> INFO [2020-10-20 22:58:42] Iter 7 complete. delta = 0.00305971137682195, alpha = 119.373289770871
#> INFO [2020-10-20 22:58:42] Iter 8 complete. delta = 0.00223438531278186, alpha = 110.751893001231
#> INFO [2020-10-20 22:58:43] Iter 9 complete. delta = 0.00169113161713632, alpha = 103.79448253191
#> INFO [2020-10-20 22:58:43] Iter 10 complete. delta = 0.00131704565908882, alpha = 98.0666848277102
#> INFO [2020-10-20 22:58:43] Iter 11 complete. delta = 0.00104990445694986, alpha = 93.2724628409278
#> INFO [2020-10-20 22:58:44] Iter 12 complete. delta = 0.000853327186998111, alpha = 89.2032943633571
#> INFO [2020-10-20 22:58:44] Iter 13 complete. delta = 0.00070499175718764, alpha = 85.708170148932
#> INFO [2020-10-20 22:58:44] Iter 14 complete. delta = 0.000590640643901349, alpha = 82.6751084781295
#> INFO [2020-10-20 22:58:44] Iter 15 complete. delta = 0.000500851850425585, alpha = 80.0193484656878
#> INFO [2020-10-20 22:58:44] Iter 16 complete. delta = 0.000429213933880499, alpha = 77.6755734092277
#> INFO [2020-10-20 22:58:45] Iter 17 complete. delta = 0.000371251575275382, alpha = 75.59264977873
#> INFO [2020-10-20 22:58:45] Iter 18 complete. delta = 0.000323770320750219, alpha = 73.7299833766569
#> INFO [2020-10-20 22:58:45] Iter 19 complete. delta = 0.000284444461302715, alpha = 72.0549422875352
#> INFO [2020-10-20 22:58:45] Iter 20 complete. delta = 0.000251550691895036, alpha = 70.5409998794925

#> Warning: 
#> Reached maximum allowed iterations. Returning early.

# build a rank-5 approximation only for
# observed elements of ml100k

preds <- predict(mf, ml100k)

# estimate the in-sample reconstruction mse

R <- resid(mf, ml100k)
norm(R, type = "F") / nnzero(ml100k)

#> [1] 0.003424297

### Exact AdaptiveInitialize initialization ---------------------------------

mf2 <- adaptive_impute(
  ml100k,
  rank = 3L,
  max_iter = 20L,
  initialization = "adaptive-initialize"
)

#> INFO [2020-10-20 22:58:45] Using adaptive-initialize initialization.
#> INFO [2020-10-20 22:58:45] Beginning AdaptiveInitialize.
#> INFO [2020-10-20 22:58:45] p_hat = 0.0630466936422453, non-zero entries = 100000, total entries = 1586126
#> INFO [2020-10-20 22:58:47] Calculating nuclear norm (slow step). Using exact method.
#> INFO [2020-10-20 22:59:04] Computation complete, alpha = 684.448307778619
#> INFO [2020-10-20 22:59:04] lambda_hat = 10108.0189264858, 3763.72941447119, 3334.41923001022
#> INFO [2020-10-20 22:59:04] Done initializing.
#> INFO [2020-10-20 22:59:04] Beginning AdaptiveImpute (max 20 iterations).
#> INFO [2020-10-20 22:59:04] Checking convergence every 1 iteration(s).
#> INFO [2020-10-20 22:59:04] Iter 1 complete. delta = 0.331729005563784, alpha = 9739.2282893819
#> INFO [2020-10-20 22:59:04] Iter 2 complete. delta = 0.0862220431883504, alpha = 1935.70067515593
#> INFO [2020-10-20 22:59:04] Iter 3 complete. delta = 0.0294556910363757, alpha = 714.201342288225
#> INFO [2020-10-20 22:59:05] Iter 4 complete. delta = 0.0125670215501749, alpha = 365.345674475765
#> INFO [2020-10-20 22:59:05] Iter 5 complete. delta = 0.00626715789948975, alpha = 229.111119173539
#> INFO [2020-10-20 22:59:05] Iter 6 complete. delta = 0.00350123681544133, alpha = 164.520680178944
#> INFO [2020-10-20 22:59:05] Iter 7 complete. delta = 0.00213131597713726, alpha = 129.54668546913
#> INFO [2020-10-20 22:59:06] Iter 8 complete. delta = 0.00138673542316825, alpha = 108.688760352002
#> INFO [2020-10-20 22:59:06] Iter 9 complete. delta = 0.000951113055993692, alpha = 95.306982997116
#> INFO [2020-10-20 22:59:06] Iter 10 complete. delta = 0.000680657489971074, alpha = 86.2193491031794
#> INFO [2020-10-20 22:59:06] Iter 11 complete. delta = 0.000504398368615498, alpha = 79.7614772171736
#> INFO [2020-10-20 22:59:07] Iter 12 complete. delta = 0.000384826841115098, alpha = 74.9995043987432
#> INFO [2020-10-20 22:59:07] Iter 13 complete. delta = 0.00030094382911392, alpha = 71.3785131501534
#> INFO [2020-10-20 22:59:07] Iter 14 complete. delta = 0.000240402715149437, alpha = 68.5526904885565
#> INFO [2020-10-20 22:59:07] Iter 15 complete. delta = 0.000195632567899946, alpha = 66.2977512249919
#> INFO [2020-10-20 22:59:07] Iter 16 complete. delta = 0.000161819505310703, alpha = 64.4632187901157
#> INFO [2020-10-20 22:59:08] Iter 17 complete. delta = 0.000135805018725351, alpha = 62.9451652703282
#> INFO [2020-10-20 22:59:08] Iter 18 complete. delta = 0.00011545931197138, alpha = 61.6699892590871
#> INFO [2020-10-20 22:59:08] Iter 19 complete. delta = 9.93115075385744e-05, alpha = 60.5844126227797
#> INFO [2020-10-20 22:59:08] Iter 20 complete. delta = 8.63241846551432e-05, alpha = 59.6491171377176

#> Warning: 
#> Reached maximum allowed iterations. Returning early.

R2 <- resid(mf2, ml100k)
norm(R2, type = "F") / nnzero(ml100k)

#> [1] 0.003153474

### Approximate AdaptiveInitialize initialization ---------------------------

mf3 <- adaptive_impute(
  ml100k,
  rank = 3L,
  max_iter = 20L,
  initialization = "approximate",
  additional = 25
)

#> INFO [2020-10-20 22:59:08] Using approximate initialization.
#> INFO [2020-10-20 22:59:08] Beginning AdaptiveInitialize.
#> INFO [2020-10-20 22:59:08] p_hat = 0.0630466936422453, non-zero entries = 100000, total entries = 1586126
#> INFO [2020-10-20 22:59:09] Calculating nuclear norm (slow step). Using approximate method.
#> INFO [2020-10-20 22:59:09] Approximating alpha using rank 28 approximation.
#> INFO [2020-10-20 22:59:09] Computation complete, alpha = 122.705383069511
#> INFO [2020-10-20 22:59:09] lambda_hat = 10115.0071508113, 3782.45715747362, 3355.54386712941
#> INFO [2020-10-20 22:59:09] Done initializing.
#> INFO [2020-10-20 22:59:09] Beginning AdaptiveImpute (max 20 iterations).
#> INFO [2020-10-20 22:59:09] Checking convergence every 1 iteration(s).
#> INFO [2020-10-20 22:59:10] Iter 1 complete. delta = 0.331804512800994, alpha = 9778.28449154083
#> INFO [2020-10-20 22:59:10] Iter 2 complete. delta = 0.086323076286072, alpha = 1944.24065799898
#> INFO [2020-10-20 22:59:10] Iter 3 complete. delta = 0.0295131978223363, alpha = 717.435048499764
#> INFO [2020-10-20 22:59:10] Iter 4 complete. delta = 0.0125994031576986, alpha = 366.952789446032
#> INFO [2020-10-20 22:59:11] Iter 5 complete. delta = 0.00628639973387419, alpha = 230.049797656835
#> INFO [2020-10-20 22:59:11] Iter 6 complete. delta = 0.00351339430048935, alpha = 165.129973513147
#> INFO [2020-10-20 22:59:11] Iter 7 complete. delta = 0.00213944370268515, alpha = 129.971842279086
#> INFO [2020-10-20 22:59:11] Iter 8 complete. delta = 0.00139242928770028, alpha = 109.001043699088
#> INFO [2020-10-20 22:59:12] Iter 9 complete. delta = 0.000955249970252651, alpha = 95.5451888833827
#> INFO [2020-10-20 22:59:12] Iter 10 complete. delta = 0.000683747680615078, alpha = 86.4064022472185
#> INFO [2020-10-20 22:59:12] Iter 11 complete. delta = 0.000506755978866302, alpha = 79.9118290393169
#> INFO [2020-10-20 22:59:12] Iter 12 complete. delta = 0.000386655343914966, alpha = 75.1227397999686
#> INFO [2020-10-20 22:59:12] Iter 13 complete. delta = 0.000302380751005925, alpha = 71.4812480906727
#> INFO [2020-10-20 22:59:13] Iter 14 complete. delta = 0.000241544267943076, alpha = 68.6396377493615
#> INFO [2020-10-20 22:59:13] Iter 15 complete. delta = 0.00019654792785005, alpha = 66.3723533678558
#> INFO [2020-10-20 22:59:13] Iter 16 complete. delta = 0.000162559501467102, alpha = 64.5280422949026
#> INFO [2020-10-20 22:59:13] Iter 17 complete. delta = 0.000136407653922469, alpha = 63.0021559218108
#> INFO [2020-10-20 22:59:14] Iter 18 complete. delta = 0.000115953404751415, alpha = 61.7206432289173
#> INFO [2020-10-20 22:59:14] Iter 19 complete. delta = 9.97191709473208e-05, alpha = 60.6298943105331
#> INFO [2020-10-20 22:59:14] Iter 20 complete. delta = 8.66625574651726e-05, alpha = 59.6903422598712

#> Warning: 
#> Reached maximum allowed iterations. Returning early.

R3 <- resid(mf3, ml100k)
norm(R3, type = "F") / nnzero(ml100k)

#> [1] 0.003154522