AdaptiveInitialize

An implementation of the AdaptiveInitialize algorithm for matrix imputation for sparse matrices. At the moment the implementation is only suitable for small matrices with on the order of thousands of rows and columns at most.

adaptive_initialize(
  X,
  rank,
  ...,
  p_hat = NULL,
  alpha_method = c("exact", "approximate"),
  additional = NULL
)

# S3 method for sparseMatrix
adaptive_initialize(
  X,
  rank,
  ...,
  p_hat = NULL,
  alpha_method = c("exact", "approximate"),
  additional = NULL
)

Arguments

X
A sparse matrix of sparseMatrix class. Explicit (observed) zeroes in X can be dropped for

rank
Desired rank (integer) to use in the low rank approximation. Must be at least 2L and at most the rank of X.

...
Ignored.

p_hat
The portion of X that is observed. Defaults to NULL, in which case p_hat is set to the number of observed elements of X. Primarily for internal use in citation_impute() or advanced users.

alpha_method
Either "exact" or "approximate", defaulting to "exact". "exact" is computationally expensive and requires taking a complete SVD of matrix of size nrow(X) x nrow(X), and matches the AdaptiveInitialize algorithm exactly. "approximate" departs from the AdaptiveInitialization algorithm to compute a truncated SVD of rank rank + additional instead of a complete SVD. This reduces computational burden, but the resulting estimates of singular-ish values will not be penalized as much as in the AdaptiveInitialize algorithm.

additional
Ignored except when alpha_method = "approximate" in which case it controls the precise of the approximation to alpha. The approximate computation of alpha will always understand alpha, but the approximation will be better for larger values of additional. We recommend making additional as large as computationally tolerable.

Value

A low rank matrix factorization represented by an adaptive_imputation() object.

Examples

mf <- adaptive_initialize(ml100k, rank = 3L)

#> INFO [2020-10-21 02:07:48] Beginning AdaptiveInitialize.
#> INFO [2020-10-21 02:07:48] p_hat = 0.0630466936422453, non-zero entries = 100000, total entries = 1586126
#> INFO [2020-10-21 02:07:48] Calculating nuclear norm (slow step). Using exact method.
#> INFO [2020-10-21 02:08:05] Computation complete, alpha = 684.448307778619
#> INFO [2020-10-21 02:08:05] lambda_hat = 10108.0189264858, 3763.72941447119, 3334.41923001022

mf

#> 
#> Adaptively Imputed Low Rank Matrix Factorization
#> ------------------------------------------------
#> 
#> Rank: 3
#> 
#> Rows: 943
#> Cols: 1682
#> 
#> d[rank]: 3334.419
#> alpha:   684.448
#> 
#> Components
#> 
#> u: 943 x 3 [matrix] 
#> d: 3      [numeric] 
#> v: 1682 x 3 [matrix] 

# build a rank-5 approximation only for
# observed elements of ml100k
preds <- predict(mf, ml100k)