CitationImpute

An implementation of the AdaptiveImpute algorithm using efficient sparse matrix computations, specialized for the case when missing values in the upper triangle are taken to be explicitly observed zeros, as opposed to missing values. This is primarily useful for spectral decompositions of adjacency matrices of graphs with (near) tree structure, such as citation networks.

citation_impute(
  X,
  rank,
  ...,
  initialization = c("svd", "adaptive-initialize", "approximate"),
  max_iter = 200L,
  check_interval = 1L,
  epsilon = 1e-07,
  additional = NULL
)

# S3 method for sparseMatrix
citation_impute(
  X,
  rank,
  ...,
  initialization = c("svd", "adaptive-initialize", "approximate"),
  additional = NULL
)

# S3 method for LRMF
citation_impute(
  X,
  rank,
  ...,
  epsilon = 1e-07,
  max_iter = 200L,
  check_interval = 1L
)

Arguments

X
A square sparse matrix of Matrix::sparseMatrix() class. Implicit zeros in the upper triangle of this matrix are considered observed and predictions on these elements contribute to the objective function minimized by AdaptiveImpute.

rank
Desired rank (integer) to use in the low rank approximation. Must be at least 2L and at most the rank of X. Note that the rank of X is typically unobserved and computations may be unstable or even fail when rank is near or exceeds this threshold.

...
Unused additional arguments.

initialization
How to initialize the low rank approximation. Options are:

"svd" (default). In the initialization step, this treats unobserved values as zeroes.

"adaptive-initialize". In the initialization step, this treats unobserved values as actually unobserved. However, the current AdaptiveInitialize implementation relies on dense matrix computations that are only suitable for relatively small matrices.

"approximate". An approximate variant of AdaptiveInitialize that is less computationally expensive. See adaptive_initialize for details.

Note that initialization matters as AdaptiveImpute optimizes a non-convex objective. The current theory shows that initializing with AdaptiveInitialize leads to a consistent estimator, but it isn't know if this is the case for SVD initialization. Empirically we have found that SVD initialization works well nonetheless.

max_iter
Maximum number of iterations to perform (integer). Defaults to 200L. In practice 10 or so iterations will get you a decent approximation to use in exploratory analysis, and and 50-100 will get you most of the way to convergence. Must be at least 1L.

check_interval
Integer specifying how often to perform convergence checks. Defaults to 1L. In practice, check for convergence requires a norm calculation that is expensive for large matrices and decreasing the frequency of convergence checks will reduce computation time. Can also be set to NULL, which case max_iter iterations of the algorithm will occur with no possibility of stopping due to small relative change in the imputed matrix. In this case delta will be reported as Inf.

epsilon
Convergence criteria, measured in terms of relative change in Frobenius norm of the full imputed matrix. Defaults to 1e-7.

additional
Ignored except when alpha_method = "approximate" in which case it controls the precise of the approximation to alpha. The approximate computation of alpha will always understand alpha, but the approximation will be better for larger values of additional. We recommend making additional as large as computationally tolerable.

Value

A low rank matrix factorization represented by an adaptive_imputation() object.

Examples

# create a (binary) square sparse matrix to demonstrate on

n <- 100
A <- rsparsematrix(n, n, 0.1, rand.x = NULL)

### SVD initialization (default) --------------------------------------------

mf <- citation_impute(A, rank = 3L, max_iter = 10L)

#> INFO [2020-10-21 02:08:05] Matrix has 1000 non-zero elements, 510 in the upper triangle (including the diagonal), and 490 in the strict lower triangle.
#> INFO [2020-10-21 02:08:05] Using svd initialization.
#> INFO [2020-10-21 02:08:05] Done initializing.
#> INFO [2020-10-21 02:08:05] Beginning AdaptiveImpute (max 10 iterations).
#> INFO [2020-10-21 02:08:05] Checking convergence every 1 iteration(s).
#> INFO [2020-10-21 02:08:05] Iter 1 complete. delta = 0.214455844482967, alpha = 7.19872761929092
#> INFO [2020-10-21 02:08:05] Iter 2 complete. delta = 0.0655816579961252, alpha = 6.64450388191785
#> INFO [2020-10-21 02:08:06] Iter 3 complete. delta = 0.0282828497742711, alpha = 6.29817854147802
#> INFO [2020-10-21 02:08:06] Iter 4 complete. delta = 0.0149131921561725, alpha = 6.06364670460596
#> INFO [2020-10-21 02:08:06] Iter 5 complete. delta = 0.00890970590501811, alpha = 5.89548036092589
#> INFO [2020-10-21 02:08:06] Iter 6 complete. delta = 0.00579959384483573, alpha = 5.76959308559485
#> INFO [2020-10-21 02:08:06] Iter 7 complete. delta = 0.0040222915648417, alpha = 5.67204772446829
#> INFO [2020-10-21 02:08:07] Iter 8 complete. delta = 0.00293049425043196, alpha = 5.59426170122
#> INFO [2020-10-21 02:08:07] Iter 9 complete. delta = 0.00222093116373994, alpha = 5.5307039336318
#> INFO [2020-10-21 02:08:07] Iter 10 complete. delta = 0.00173822964715284, alpha = 5.47768100920203

#> Warning: 
#> Reached maximum allowed iterations. Returning early.

### Exact AdaptiveInitialize initialization ---------------------------------

mf2 <- citation_impute(
  A,
  rank = 3L,
  max_iter = 10L,
  initialization = "adaptive-initialize"
)

#> INFO [2020-10-21 02:08:07] Matrix has 1000 non-zero elements, 510 in the upper triangle (including the diagonal), and 490 in the strict lower triangle.
#> INFO [2020-10-21 02:08:07] Using adaptive-initialize initialization.
#> INFO [2020-10-21 02:08:07] Beginning AdaptiveInitialize.
#> INFO [2020-10-21 02:08:07] p_hat = 0.5341, non-zero entries = 1000, total entries = 10000
#> INFO [2020-10-21 02:08:07] Calculating nuclear norm (slow step). Using exact method.
#> INFO [2020-10-21 02:08:07] Computation complete, alpha = 6.43272360063481
#> INFO [2020-10-21 02:08:07] lambda_hat = 19.5542188962643, 8.86806066342759, 8.6815774967917
#> INFO [2020-10-21 02:08:07] Done initializing.
#> INFO [2020-10-21 02:08:07] Beginning AdaptiveImpute (max 10 iterations).
#> INFO [2020-10-21 02:08:07] Checking convergence every 1 iteration(s).
#> INFO [2020-10-21 02:08:07] Iter 1 complete. delta = 0.107094385252634, alpha = 6.66993591988108
#> INFO [2020-10-21 02:08:08] Iter 2 complete. delta = 0.0279517637430547, alpha = 6.30339393286254
#> INFO [2020-10-21 02:08:08] Iter 3 complete. delta = 0.0150626880745451, alpha = 6.06584387384967
#> INFO [2020-10-21 02:08:08] Iter 4 complete. delta = 0.00905399018631845, alpha = 5.89497466377753
#> INFO [2020-10-21 02:08:08] Iter 5 complete. delta = 0.00589202079182546, alpha = 5.76696989181021
#> INFO [2020-10-21 02:08:08] Iter 6 complete. delta = 0.00407714705762582, alpha = 5.66788203108073
#> INFO [2020-10-21 02:08:09] Iter 7 complete. delta = 0.00296236276281234, alpha = 5.58899995835211
#> INFO [2020-10-21 02:08:09] Iter 8 complete. delta = 0.00223929627619432, alpha = 5.524665682593
#> INFO [2020-10-21 02:08:09] Iter 9 complete. delta = 0.00174875795612436, alpha = 5.47108714032427
#> INFO [2020-10-21 02:08:09] Iter 10 complete. delta = 0.00140314681717488, alpha = 5.42565281799271

#> Warning: 
#> Reached maximum allowed iterations. Returning early.

### Approximate AdaptiveInitialize initialization ---------------------------

mf3 <- citation_impute(
  A,
  rank = 3L,
  max_iter = 10L,
  initialization = "approximate",
  additional = 5L
)

#> INFO [2020-10-21 02:08:09] Matrix has 1000 non-zero elements, 510 in the upper triangle (including the diagonal), and 490 in the strict lower triangle.
#> INFO [2020-10-21 02:08:09] Using approximate initialization.
#> INFO [2020-10-21 02:08:09] Beginning AdaptiveInitialize.
#> INFO [2020-10-21 02:08:09] p_hat = 0.5341, non-zero entries = 1000, total entries = 10000
#> INFO [2020-10-21 02:08:09] Calculating nuclear norm (slow step). Using approximate method.
#> INFO [2020-10-21 02:08:09] Approximating alpha using rank 8 approximation.
#> INFO [2020-10-21 02:08:09] Computation complete, alpha = 1.18127864248873
#> INFO [2020-10-21 02:08:09] lambda_hat = 20.0194061721633, 9.85147942053702, 9.68395243048171
#> INFO [2020-10-21 02:08:09] Done initializing.
#> INFO [2020-10-21 02:08:09] Beginning AdaptiveImpute (max 10 iterations).
#> INFO [2020-10-21 02:08:09] Checking convergence every 1 iteration(s).
#> INFO [2020-10-21 02:08:10] Iter 1 complete. delta = 0.111048751408125, alpha = 6.63225678210305
#> INFO [2020-10-21 02:08:10] Iter 2 complete. delta = 0.0257212887841839, alpha = 6.27770401359615
#> INFO [2020-10-21 02:08:10] Iter 3 complete. delta = 0.0141298120405171, alpha = 6.05039979385688
#> INFO [2020-10-21 02:08:10] Iter 4 complete. delta = 0.00861674757582726, alpha = 5.88577337685571
#> INFO [2020-10-21 02:08:10] Iter 5 complete. delta = 0.00566189321294998, alpha = 5.76178252193578
#> INFO [2020-10-21 02:08:11] Iter 6 complete. delta = 0.00394538779805784, alpha = 5.66540633961094
#> INFO [2020-10-21 02:08:11] Iter 7 complete. delta = 0.00288225329574759, alpha = 5.58842506392507
#> INFO [2020-10-21 02:08:11] Iter 8 complete. delta = 0.00218847687103855, alpha = 5.52545993076456
#> INFO [2020-10-21 02:08:11] Iter 9 complete. delta = 0.00171557083514344, alpha = 5.47288743772458
#> INFO [2020-10-21 02:08:11] Iter 10 complete. delta = 0.00138108359531644, alpha = 5.42820297945485

#> Warning: 
#> Reached maximum allowed iterations. Returning early.