AdaptiveImpute

An implementation of the AdaptiveImpute algorithm for matrix completion for sparse matrices.

adaptive_impute(
  X,
  rank,
  ...,
  initialization = c("svd", "adaptive-initialize"),
  max_iter = 200L,
  check_interval = 1L,
  epsilon = 1e-07
)

# S3 method for sparseMatrix
adaptive_impute(X, rank, initialization = c("svd", "adaptive-initialize"), ...)

# S3 method for LRMF
adaptive_impute(
  mf,
  X,
  ...,
  epsilon = 1e-07,
  max_iter = 200L,
  check_interval = 1L,
  verbose = FALSE
)

# S3 method for sparseMatrix
adaptive_initialize(X, rank, p_hat = NULL, ...)

Arguments

X
A sparse matrix of Matrix::sparseMatrix() class.

rank
Desired rank (integer) to use in the low rank approximation.

...
Unused additional arguments.

initialization
How to initialize the low rank approximation. Options are:

"svd" (default). In the initialization step, this treats unobserved values as zeroes.

"adaptive-initialize". In the initialization step, this treats unobserved values as actually unobserved. However, the current AdaptiveInitialize implementation relies on dense matrix computations that are only suitable for relatively small matrices.

Note that initialization matters as AdaptiveImpute optimizes a non-convex objective. The current theory shows that initializing with AdaptiveInitialize leads to a consistent estimator, but it isn't know if this is the case for SVD initialization. Empirically we have found that SVD initialization works well nonetheless.

max_iter
Maximum number of iterations to perform (integer). Defaults to 200L. In practice 10 or so iterations will get you a decent approximation to use in exploratory analysis, and and 50-100 will get you most of the way to convergence.

check_interval
Integer specifying how often to perform convergence checks. Defaults to 1L. In practice, check for convergence requires a norm calculation that is expensive for large matrices and decreasing the frequency of convergence checks will reduce computation time.

epsilon
Convergence criteria, measured in terms of relative change in Frobenius norm of the full imputed matrix. Defaults to 1e-7.

Value

A low rank matrix factorization represented by an LRMF object. See LRMF3::mf() for details.

Examples

mf <- adaptive_impute(ml100k, rank = 3L, max_iter = 20L)

#> INFO [2020-09-22 17:09:37] Use svd initialization.
#> INFO [2020-09-22 17:09:37] Done initializing.
#> INFO [2020-09-22 17:09:37] Beginning AdaptiveImpute (max 20 iterations).
#> INFO [2020-09-22 17:09:37] Checking convergence every 1 iteration(s).
#> INFO [2020-09-22 17:09:37] Iter 1 complete. delta = 0.22312337, alpha = 184.71
#> INFO [2020-09-22 17:09:38] Iter 2 complete. delta = 0.05159124, alpha = 154.411
#> INFO [2020-09-22 17:09:38] Iter 3 complete. delta = 0.02125919, alpha = 135.61
#> INFO [2020-09-22 17:09:39] Iter 4 complete. delta = 0.01114668, alpha = 122.308
#> INFO [2020-09-22 17:09:39] Iter 5 complete. delta = 0.00669206, alpha = 112.354
#> INFO [2020-09-22 17:09:39] Iter 6 complete. delta = 0.00438637, alpha = 104.612
#> INFO [2020-09-22 17:09:40] Iter 7 complete. delta = 0.00305875, alpha = 98.411
#> INFO [2020-09-22 17:09:40] Iter 8 complete. delta = 0.00223373, alpha = 93.328
#> INFO [2020-09-22 17:09:41] Iter 9 complete. delta = 0.00169067, alpha = 89.081
#> INFO [2020-09-22 17:09:41] Iter 10 complete. delta = 0.00131671, alpha = 85.477
#> INFO [2020-09-22 17:09:41] Iter 11 complete. delta = 0.00104966, alpha = 82.378
#> INFO [2020-09-22 17:09:42] Iter 12 complete. delta = 0.00085314, alpha = 79.684
#> INFO [2020-09-22 17:09:42] Iter 13 complete. delta = 0.00070485, alpha = 77.319
#> INFO [2020-09-22 17:09:43] Iter 14 complete. delta = 0.00059053, alpha = 75.227
#> INFO [2020-09-22 17:09:43] Iter 15 complete. delta = 0.00050077, alpha = 73.362
#> INFO [2020-09-22 17:09:44] Iter 16 complete. delta = 0.00042915, alpha = 71.69
#> INFO [2020-09-22 17:09:44] Iter 17 complete. delta = 0.0003712, alpha = 70.182
#> INFO [2020-09-22 17:09:44] Iter 18 complete. delta = 0.00032373, alpha = 68.815
#> INFO [2020-09-22 17:09:45] Iter 19 complete. delta = 0.00028441, alpha = 67.571
#> INFO [2020-09-22 17:09:45] Iter 20 complete. delta = 0.00025153, alpha = 66.434

#> Warning: 
#> Reached maximum allowed iterations. Returning early.

# mf

# build a rank-5 approximation only for
# observed elements of ml100k

preds <- predict(mf, ml100k)

# estimate the in-sample reconstruction mse

R <- resid(mf, ml100k)
norm(R, type = "F") / nnzero(ml100k)

#> [1] 0.003424301

mf2 <- adaptive_impute(
  ml100k,
  rank = 3L,
  max_iter = 20L,
  initialization = "adaptive-initialize"
)

#> INFO [2020-09-22 17:09:45] Use adaptive-initialize initialization.
#> INFO [2020-09-22 17:09:45] Beginning AdaptiveInitialize.
#> INFO [2020-09-22 17:09:45] p_hat = 0.0630466936422453, non-zero entries = 100000, total entries = 1586126
#> INFO [2020-09-22 17:09:48] Starting nuclear norm computation (slow step).
#> INFO [2020-09-22 17:10:10] Computation complete, alpha = 684.44830777862
#> INFO [2020-09-22 17:10:10] lambda_hat = 10108.0189264858, 3763.72941447119, 3334.41923001022
#> INFO [2020-09-22 17:10:10] Done initializing.
#> INFO [2020-09-22 17:10:10] Beginning AdaptiveImpute (max 20 iterations).
#> INFO [2020-09-22 17:10:10] Checking convergence every 1 iteration(s).
#> INFO [2020-09-22 17:10:11] Iter 1 complete. delta = 0.33221539, alpha = 47012.206
#> INFO [2020-09-22 17:10:11] Iter 2 complete. delta = 0.08622896, alpha = 6190.828
#> INFO [2020-09-22 17:10:11] Iter 3 complete. delta = 0.02941825, alpha = 1948.79
#> INFO [2020-09-22 17:10:12] Iter 4 complete. delta = 0.0125374, alpha = 884.925
#> INFO [2020-09-22 17:10:12] Iter 5 complete. delta = 0.00624713, alpha = 494.425
#> INFO [2020-09-22 17:10:13] Iter 6 complete. delta = 0.00348784, alpha = 316.976
#> INFO [2020-09-22 17:10:13] Iter 7 complete. delta = 0.00212217, alpha = 224.488
#> INFO [2020-09-22 17:10:13] Iter 8 complete. delta = 0.00138032, alpha = 171.421
#> INFO [2020-09-22 17:10:14] Iter 9 complete. delta = 0.00094648, alpha = 138.71
#> INFO [2020-09-22 17:10:14] Iter 10 complete. delta = 0.00067723, alpha = 117.383
#> INFO [2020-09-22 17:10:14] Iter 11 complete. delta = 0.00050181, alpha = 102.832
#> INFO [2020-09-22 17:10:15] Iter 12 complete. delta = 0.00038284, alpha = 92.519
#> INFO [2020-09-22 17:10:15] Iter 13 complete. delta = 0.00029939, alpha = 84.972
#> INFO [2020-09-22 17:10:16] Iter 14 complete. delta = 0.00023918, alpha = 79.294
#> INFO [2020-09-22 17:10:16] Iter 15 complete. delta = 0.00019465, alpha = 74.917
#> INFO [2020-09-22 17:10:16] Iter 16 complete. delta = 0.00016103, alpha = 71.471
#> INFO [2020-09-22 17:10:17] Iter 17 complete. delta = 0.00013516, alpha = 68.707
#> INFO [2020-09-22 17:10:17] Iter 18 complete. delta = 0.00011493, alpha = 66.454
#> INFO [2020-09-22 17:10:17] Iter 19 complete. delta = 9.887e-05, alpha = 64.588
#> INFO [2020-09-22 17:10:18] Iter 20 complete. delta = 8.596e-05, alpha = 63.024

#> Warning: 
#> Reached maximum allowed iterations. Returning early.

# mf2

R2 <- resid(mf2, ml100k)
norm(R2, type = "F") / nnzero(ml100k)

#> [1] 0.0031521