AdaptiveInitialize

An implementation of the AdaptiveInitialize algorithm for matrix imputation for sparse matrices. At the moment the implementation is only suitable for small matrices with on the order of thousands of rows and columns at most.

adaptive_initialize(X, rank, p_hat = NULL, ...)

Arguments

X
A sparse matrix of sparseMatrix class. Explicit (observed) zeroes in X can be dropped for

rank
Desired rank (integer) to use in the low rank approximation.

p_hat
The portion of X that is observed. Defaults to NULL, in which case p_hat is set to the number of observed elements of X. Useful when you have many explicit zeros that w

...
Ignored.

Value

A low rank matrix factorization represented by an LRMF object. See LRMF3::mf() for details.

Examples

mf <- adaptive_initialize(ml100k, rank = 3L)

#> INFO [2020-09-22 17:10:19] Beginning AdaptiveInitialize.
#> INFO [2020-09-22 17:10:19] p_hat = 0.0630466936422453, non-zero entries = 100000, total entries = 1586126
#> INFO [2020-09-22 17:10:20] Starting nuclear norm computation (slow step).
#> INFO [2020-09-22 17:10:42] Computation complete, alpha = 684.44830777862
#> INFO [2020-09-22 17:10:42] lambda_hat = 10108.0189264858, 3763.72941447119, 3334.41923001022

mf

#> 
#> Adaptively Imputed Low Rank Matrix Factorization
#> ------------------------------------------------
#> 
#> Rank: 3
#> 
#> Rows: 943
#> Cols: 1682
#> 
#> d[rank]: 3334.419
#> alpha:   684.448
#> 
#> Components
#> 
#> u: 943 x 3 [matrix] 
#> d: 3      [numeric] 
#> v: 1682 x 3 [matrix] 

# build a rank-5 approximation only for
# observed elements of ml100k
preds <- predict(mf, ml100k)