fastadi: Self-Tuning Data Adaptive Matrix Imputation

Implements the AdaptiveImpute matrix completion algorithm of 'Intelligent Initialization and Adaptive Thresholding for Iterative Matrix Completion' doi:10.1080/10618600.2018.1518238 as well as the specialized variant of 'Co-Factor Analysis of Citation Networks' doi:10.1080/10618600.2024.2394464 . AdaptiveImpute is useful for embedding sparsely observed matrices, often out performs competing matrix completion algorithms, and self-tunes its hyperparameter, making usage easy.

See also

Author

Maintainer: Alex Hayes alexpghayes@gmail.com (ORCID) [copyright holder]

Authors:

Juhee Cho

Donggyu Kim

Karl Rohe