Expand an SVD only at observed values of a sparse matrix

TODO: describe what it looks like for dimensions to match up between s and mask. See vignette("sparse-computations") for mathematical details.

Usage

masked_approximation_impl(U, V, row, col)

Arguments

U

Low-rank matrix of left singular-ish vectors.

V

Low-rank matrix of right singular-ish vectors.

row

Zero-based row indices of observed elements.

col

Zero-based col indices of observed elements.

Value

A sparse matrix representing the low-rank reconstruction from U, d and V, only at the index pairs indicated by row and col.

Details

The idea is to populate U, d and V with using the elements of an SVD-like list. You can generate row and col most easily from a sparse masking Matrix (Matrix package), coercing to triplet format, and extracting mask@i for row and mask@j for column.